+++ financeager/cli.py
@@ -12,7 +12,7 @@ import pkg_resources
 
 import financeager
 
-from . import (PERIOD_DATE_FORMAT, __version__, clients, config, entries,
+from . import (POCKET_DATE_FORMAT, __version__, clients, config, entries,
                exceptions, init_logger, listing,
                make_log_stream_handler_verbose, setup_log_file_handler)
 
@@ -138,7 +138,7 @@ def _preprocess(data, date_format=None):
     # has already been converted
     if date is not None and date_format is not None:
         try:
-            date = time.strftime(PERIOD_DATE_FORMAT,
+            date = time.strftime(POCKET_DATE_FORMAT,
                                  time.strptime(date, date_format))
             data["date"] = date
         except ValueError:
@@ -202,7 +202,7 @@ def _format_response(response, command, **listing_options):
     """Format the given response (dict or str) into human-readable text.
     If the response is a string, it is immediately returned.
     If the response does not contain any of the fields 'id', 'elements',
-    'element', or 'periods', an empty string is returned.
+    'element', or 'pockets', an empty string is returned.
     The 'listing_options' are passed to listing.prettify().
 
     :return: str
@@ -229,8 +229,8 @@ def _format_response(response, command, **listing_options):
         return entries.prettify(
             element, default_category=listing_options["default_category"])
 
-    periods = response.get("periods", [])
-    return "\n".join([p for p in periods])
+    pockets = response.get("pockets", [])
+    return "\n".join([p for p in pockets])
 
 
 def _parse_command(args=None):
@@ -318,20 +318,20 @@ least a frequency, start date and end date are optional. Default:
         "-e", "--end", help="new end date (for recurrent entries only)")
 
     copy_parser = subparsers.add_parser(
-        "copy", help="copy an entry from one period to another")
+        "copy", help="copy an entry from one pocket to another")
     copy_parser.add_argument("eid", help="entry ID")
     copy_parser.add_argument(
         "-s",
         "--source",
         default=None,
-        dest="source_period",
-        help="period to copy the entry from")
+        dest="source_pocket",
+        help="pocket to copy the entry from")
     copy_parser.add_argument(
         "-d",
         "--destination",
         default=None,
-        dest="destination_period",
-        help="period to copy the entry to")
+        dest="destination_pocket",
+        help="pocket to copy the entry to")
     copy_parser.add_argument(
         "-t",
         "--table-name",
@@ -339,7 +339,7 @@ least a frequency, start date and end date are optional. Default:
         help="Table to copy the entry from/to. Default: 'standard'.")
 
     list_parser = subparsers.add_parser(
-        "list", help="list all entries in the period database")
+        "list", help="list all entries in the pocket database")
     list_parser.add_argument(
         "-f",
         "--filters",
@@ -375,8 +375,8 @@ least a frequency, start date and end date are optional. Default:
         help="show only entries of given month (default to current one)",
     )
 
-    periods_parser = subparsers.add_parser(
-        "periods", help="list all period databases")
+    pockets_parser = subparsers.add_parser(
+        "pockets", help="list all pocket databases")
 
     # Add common options to subparsers
     for subparser in subparsers.choices.values():
@@ -391,9 +391,9 @@ least a frequency, start date and end date are optional. Default:
             action="store_true",
             help="Be verbose about internal workings")
 
-        if subparser not in [periods_parser, copy_parser]:
+        if subparser not in [pockets_parser, copy_parser]:
             subparser.add_argument(
-                "-p", "--period", help="name of period to modify or query")
+                "-p", "--pocket", help="name of pocket to modify or query")
 
     argcomplete.autocomplete(parser)
     return vars(parser.parse_args(args=args))