+++ financeager/server.py
@@ -2,46 +2,46 @@
 import glob
 import os.path
 
-from . import default_period_name, init_logger, period
+from . import default_pocket_name, init_logger, pocket
 
 logger = init_logger(__name__)
 
 
 class Server:
-    """Server class holding the ``TinyDbPeriod`` databases.
+    """Server class holding the ``TinyDbPocket`` databases.
 
-    All database handling is taken care of in the underlying `TinyDbPeriod`.
-    Kwargs (f.i. storage) are passed to the TinyDbPeriod member.
+    All database handling is taken care of in the underlying `TinyDbPocket`.
+    Kwargs (f.i. storage) are passed to the TinyDbPocket member.
     """
 
     def __init__(self, **kwargs):
-        self._periods = {}
-        self._period_kwargs = kwargs
+        self._pockets = {}
+        self._pocket_kwargs = kwargs
 
     def run(self, command, **kwargs):
-        """The requested period is created if not yet present. The method of
-        `Period` corresponding to the given `command` is called. All `kwargs`
+        """The requested pocket is created if not yet present. The method of
+        `Pocket` corresponding to the given `command` is called. All `kwargs`
         are passed on. A json-like response is returned.
 
         Wrap this in a 'broad' try-except block to catch any server-side errors.
         :return: dict
-            key is one of 'id', 'element', 'elements', 'error', 'periods'
+            key is one of 'id', 'element', 'elements', 'error', 'pockets'
         """
         logger.debug("Running '{}' with {}".format(command, kwargs))
 
         try:
-            if command == "periods":
-                return {"periods": self._period_names()}
+            if command == "pockets":
+                return {"pockets": self._pocket_names()}
             elif command == "copy":
                 return {"id": self._copy_entry(**kwargs)}
             elif command == "stop":
                 # graceful shutdown, invoke closing of files
-                for pd in self._periods.values():
+                for pd in self._pockets.values():
                     pd.close()
                 return {}
             else:
-                period_name = kwargs.pop("period", None)
-                pd = self._get_period(period_name)
+                pocket_name = kwargs.pop("pocket", None)
+                pd = self._get_pocket(pocket_name)
 
                 if command == "add":
                     response = {"id": pd.add_entry(**kwargs)}
@@ -59,39 +59,39 @@ class Server:
                     }
                 return response
 
-        except period.PeriodException as e:
+        except pocket.PocketException as e:
             return {"error": str(e)}
 
-    def _get_period(self, name=None):
-        """Get the Period identified by 'name' from the Periods dictionary. If
-        the Period does not exist, it is created and returned. If 'name' is
-        None, the default period name is used as defined in __init__.py
+    def _get_pocket(self, name=None):
+        """Get the Pocket identified by 'name' from the Pockets dictionary. If
+        the Pocket does not exist, it is created and returned. If 'name' is
+        None, the default pocket name is used as defined in __init__.py
 
         :type name: str or None
-        :return: Period object
+        :return: Pocket object
         """
-        name = name or default_period_name()
+        name = name or default_pocket_name()
         try:
-            pd = self._periods[name]
+            pd = self._pockets[name]
         except KeyError:
-            logger.debug("Creating new Period '{}'".format(name))
-            pd = period.TinyDbPeriod(name, **self._period_kwargs)
-            self._periods[pd.name] = pd
+            logger.debug("Creating new Pocket '{}'".format(name))
+            pd = pocket.TinyDbPocket(name, **self._pocket_kwargs)
+            self._pockets[pd.name] = pd
 
         return pd
 
-    def _period_names(self):
-        """Return names of periods currently organized by the server.
+    def _pocket_names(self):
+        """Return names of pockets currently organized by the server.
         If persistent data storage was specified, all JSON files present in the
         'data_dir' are also taken into account.
 
         :return: list(str)
         """
-        # Avoid duplicate entries if Period already present in self._periods by
+        # Avoid duplicate entries if Pocket already present in self._pockets by
         # using sets
-        names = {p._name for p in self._periods.values()}
+        names = {p._name for p in self._pockets.values()}
 
-        data_dir = self._period_kwargs.get("data_dir")
+        data_dir = self._pocket_kwargs.get("data_dir")
         if data_dir is not None:
             names.update({
                 os.path.splitext(os.path.basename(f))[0]
@@ -100,18 +100,18 @@ class Server:
 
         return sorted(names)
 
-    def _copy_entry(self, source_period=None, destination_period=None,
+    def _copy_entry(self, source_pocket=None, destination_pocket=None,
                     **kwargs):
-        """Copy an entry (specified by ID and table_name) from the source period
-        to the destination period.
+        """Copy an entry (specified by ID and table_name) from the source pocket
+        to the destination pocket.
 
-        :type _period: str
+        :type _pocket: str
         :return: ID of copied entry
-        :raises: PeriodException if the source entry does not exist
+        :raises: PocketException if the source entry does not exist
         """
-        source_period = self._get_period(source_period)
-        entry_to_copy = source_period.get_entry(**kwargs)
+        source_pocket = self._get_pocket(source_pocket)
+        entry_to_copy = source_pocket.get_entry(**kwargs)
 
-        destination_period = self._get_period(destination_period)
-        return destination_period.add_entry(
+        destination_pocket = self._get_pocket(destination_pocket)
+        return destination_pocket.add_entry(
             table_name=kwargs.get("table_name"), **entry_to_copy)